requirements.txt
Babel==2.4.0
backports-abc==0.5
backports.ssl-match-hostname==3.5.0.1
blinker==1.4
certifi==2017.4.17
chardet==3.0.4
click==6.7
Django==1.11.2
Flask==0.9
Flask-Babel==0.8
Flask-BabelEx==0.9.3
Flask-Cache==0.11
Flask-Login==0.1.3
Flask-Mail==0.9.1
Flask-MongoKit==0.6
Flask-Script==2.0.5
Flask-WTF==0.8.3
idna==2.5
image==1.5.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
mongokit==0.9.1.1
olefile==0.44
Pillow
pip==9.0.1
pymongo==2.8
python-slugify==1.2.4
pytz==2017.2
redis==2.10.5
requests==2.18.1
setuptools==1.4.2
singledispatch==3.4.0.3
six==1.11.0
speaklater==1.3
tornado==4.5.1
Unidecode==0.4.20
urllib3==1.21.1
Werkzeug==0.8.3
wheel==0.29.0
WTForms==2.1
xlrd==1.0.0
xlwt==1.2.0
jsondiff==1.1.1
w3lib==1.19.0
uWSGI==2.0.17.1
gevent==1.3.3
grpcio==1.14.2
google-api-python-client==1.7.3
google-auth==1.5.0
google-auth-httplib2==0.0.3
google-cloud-tasks==0.2.0
google-cloud-storage==1.11.0
raven==6.9.0
paypalrestsdk==1.13.1
